adjusting pf on rpi

+++ particlefilter.h
@@ -34,7 +34,7 @@ class ParticleFilter{
 public:
     ParticleFilter();
     double lastGpsSdnM = 100;//initialization with a large value
-    void initializeParticles(int x, int y);
+    void initializeParticles(double x, double y);
     static const int PARTICLE_COUNT = 100;
     static const int GPS_DIST_ERR = 1;
     static const int radiOfEarth = 6371000;//m
